pyproject.toml
[project]
name = "inventory"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "asadaligith", email = "asad2747982@gmail.com" }
]
requires-python = ">=3.13"
dependencies = [
    "dotenv>=0.9.9",
    "openai-agents>=0.2.8",
]

[project.scripts]
inventory = "inventory:main"
main = "inventory.main:start"
todos = "inventory.todos:start"


[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"
